kernel: Fix all GCC 7.x warnings
* Features fixes picked from other commits (may..be, as I'm doing
  this partial-blindly):
  https://github.com/nathanchance/angler/commit/b7d64ea0d61147289e3c892e6792a779ad41c0b7,
  https://github.com/nathanchance/op5/commit/c78a7f8a0b86f47e2a4a13aa9ddbda517d78048f,
  https://github.com/techdro-id/stormguard_kernel/commit/170cbab1e7444c66b5139e0cc357e0d4ea567e95
* Changes from previous commit:
  - Lazily fixed all prima warnings
  - Removed -Wno flags of int-in-bool-context and format-overflow
  - No more complex comparison to fix bool-compare warning
* Warnings handled:
  - -Wbool-compare
  - -Wbool-operation
  - -Wduplicate-decl-specifier
  - -Wformat-overflow
  - -Wint-in-bool-context
  - -Wmemset-elt-size
  - -Wmisleading-indentation
  - -Wparentheses
  - -Wswitch-unreachable

[v2: Include tissot's fpc1020 & untouched prima warnings]

+++ drivers/usb/gadget/function/f_fs.c
@@ -1716,7 +1716,7 @@ static void functionfs_unbind(struct ffs_data *ffs)
 static int ffs_epfiles_create(struct ffs_data *ffs)
 {
 	struct ffs_epfile *epfile, *epfiles;
-	unsigned i, count;
+	short i, count;
 
 	ENTER();
